Fix/Revision(diag) This error should only work for typescript
Decided to revise my code and change the branches. Now it checks for if the code is TypeScript first. If so, then it checks for mistyped_strict_inequality_operator, if the code is not in TypeScript, then it will trigger typescript_non_null_assertion_not_allowed_in_javascript. This makes sense as the non-null-assertion is not allowed in JavaScript regardless of if there is a == behind it or not.